Defconst should not expand to defconstant, but only to defparameter for now.
Defconst is used on values like strings that cannot be compared with eql.

This fixes a bug when attempting to translate from linj to java, for instance
when running scripts/relinj.lisp.

When basics.lisp is compiled and then immediately loaded afterward, the error
"Unknown type bigint" is raised.
A second time where basics.lisp is not compiled anymore, but only fasl-loaded this error
disappears.

This fix avoids the error even when recompiling from scratch.