Add support for custom-import statements (sort of a quick hack).
Like java-code these allow for verbatim java to be
output. Custom-import is very similar, but for import statements. It
does not have any semantic meaning for the linj compiler.

In util.linj the statement
(custom-import "import static java.lang.Math.PI;")
allows the fix to Util.java in a62c11a5 to be made permanent when
invoking scripts/relinj.lisp.